combine inputs into one file
