+++ scss/components/components.scss
@@ -1,4 +1,5 @@
 @import "textarea_form.scss";
+@import "text_input.scss";
 @import "select_form.scss";
 @import "switch.scss";
 @import "card.scss";
