+++ scss/components/select_form.scss
@@ -1,4 +1,4 @@
-.select-form {
+div.select-form {
     @extend .form-floating;
     select {
         @extend .form-select;
