+++ scss/components/text_input.scss
@@ -1,0 +1,6 @@
+div.text-input {
+  @extend .form-floating, .mb-3;
+  input {
+    @extend .form-control;
+  }
+}
