+++ scss/components/textarea_form.scss
@@ -1,4 +1,4 @@
-.textarea-form {
+div.textarea-form {
   @extend .form-floating, .mb-3;
   textarea {
     @extend .form-control, .h-100;
