+++ src/components/mod.rs
@@ -1,4 +1,2 @@
 pub mod accordion;
-pub mod select_form;
-pub mod switch;
-pub mod textarea_form;
+pub mod inputs;
