--- src/components/switch.rs
@@ -1,30 +1,0 @@
-#![allow(non_snake_case)]
-use dioxus::prelude::*;
-
-#[inline_props]
-pub fn Switch<'a>(
-    cx: Scope<'a>,
-    label: &'a str,
-    checked: bool,
-    oninput: EventHandler<'a, bool>,
-) -> Element<'a> {
-    cx.render(rsx! {
-        div {
-            class: "switch",
-            input {
-                r#type: "checkbox",
-                id: "{label}",
-                role: "switch",
-                checked: "{checked}",
-                oninput: move |event| {
-                    let is_enabled = event.value == "true";
-                    oninput.call(is_enabled);
-                }
-            }
-            label {
-                r#for: "{label}",
-                "{label}"
-            }
-        }
-    })
-}
